Conflicts:
	TrainingMesh.v12.suo
	TrainingMesh/Grammar/Batiment/Etage.cpp
	TrainingMesh/Grammar/Batiment/Toit.cpp

+++ TrainingMesh/Grammar/Batiment/Etage.cpp
@@ -1,12 +1,12 @@
 #include "Etage.h"
 
-Etage::Etage(const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const double&)
-{
+
+Etage::Etage(const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const float&){
 
 }
 
-void Etage::G(Mesh&) const
-{
+
+void Etage::G(Mesh&) const{
 
 }
