+++ TrainingMesh/Grammar/Batiment/RDC.h
@@ -1,14 +1,15 @@
 #pragma once
 #include "Etage.h"
 #include "Toit.h"
+#include "../../Mesh/Mesh.h"
 
 
 class RDC {
 
 protected:
 	Quadrangle q;
-	double h;
+	float h;
 public:
-	RDC(const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const double&);
+	RDC(const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const Vec3<float>&, const float&);
 	void G(Mesh&) const;
 };
